% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_Utilities.R
\name{auc_obj}
\alias{auc_obj}
\title{Title Meta level Objective function: NNLS for gaussian; Rank loss for binary observations}
\usage{
auc_obj(b, X, Y)
}
\arguments{
\item{b}{Weights vector}

\item{X}{Design matrix (data frame)}

\item{Y}{Outcome variable}
}
\value{
1 - AUC
}
\description{
Title Meta level Objective function: NNLS for gaussian; Rank loss for binary observations
}
\examples{
X <- data.frame(m1 = c(0.1, 0.4, 0.8, 0.9), m2 = c(0.2, 0.3, 0.7, 0.95))
Y <- c(0, 0, 1, 1)
auc_obj(b = c(0.5, 0.5), X = X, Y = Y)
}
