% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{FranzosaE_2019_CuratedMetabolome}
\alias{FranzosaE_2019_CuratedMetabolome}
\title{Franzosa et al. 2019 training metabolome table}
\format{
A data frame with samples in rows and metabolite features in columns.
}
\source{
curatedMetagenomicData / GutMicrobiomeMetabolomeCompendium-derived tutorial data.
}
\usage{
FranzosaE_2019_CuratedMetabolome
}
\description{
Metabolomics feature table for the FranzosaE_2019 training cohort.
}
\keyword{datasets}
