% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{FranzosaE_2019_CuratedMetadata}
\alias{FranzosaE_2019_CuratedMetadata}
\title{Franzosa et al. 2019 training metadata}
\format{
A data frame with one row per sample and clinical/study covariates in columns.
}
\usage{
data(FranzosaE_2019_CuratedMetadata)
}
\description{
Sample-level metadata for the FranzosaE_2019 training cohort.
}
\keyword{datasets}
