% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{FranzosaE_2019_CuratedSpeciesProfile}
\alias{FranzosaE_2019_CuratedSpeciesProfile}
\title{Franzosa et al. 2019 training species profile}
\format{
A data frame with samples in rows and microbial species features in columns.
}
\source{
curatedMetagenomicData / GutMicrobiomeMetabolomeCompendium-derived tutorial data.
}
\usage{
FranzosaE_2019_CuratedSpeciesProfile
}
\description{
Species-level relative abundance table for the FranzosaE_2019 training cohort.
}
\keyword{datasets}
