% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{FranzosaE_2019_Validation_CuratedMetabolome}
\alias{FranzosaE_2019_Validation_CuratedMetabolome}
\title{Franzosa et al. 2019 validation metabolome table}
\format{
A data frame with samples in rows and metabolite features in columns.
}
\usage{
data(FranzosaE_2019_Validation_CuratedMetabolome)
}
\description{
Metabolomics feature table for the FranzosaE_2019 external validation cohort.
}
\keyword{datasets}
