% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{FranzosaE_2019_Validation_CuratedSpeciesProfile}
\alias{FranzosaE_2019_Validation_CuratedSpeciesProfile}
\title{Franzosa et al. 2019 validation species profile}
\format{
A data frame with samples in rows and microbial species features in columns.
}
\usage{
data(FranzosaE_2019_Validation_CuratedSpeciesProfile)
}
\description{
Species-level relative abundance table for the FranzosaE_2019 external validation cohort.
}
\keyword{datasets}
