% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegratedLearnerFromMAE.R
\name{IntegratedLearnerFromMAE}
\alias{IntegratedLearnerFromMAE}
\title{MultiAssayExperiment wrapper for IntegratedLearner}
\usage{
IntegratedLearnerFromMAE(
  mae,
  experiment,
  assay.type,
  outcome.col,
  valid.col = NULL,
  ...
)
}
\arguments{
\item{mae}{\code{\link{MultiAssayExperiment}}}

\item{experiment}{\code{Character vector} or \code{Integer vector}. Specifies
experiments from \code{mae}.}

\item{assay.type}{\code{Character vector}. Specified the name of assay for
each experiment. The length must be equal to \code{experiment}.}

\item{outcome.col}{\code{Character scalar}. Specifies the name of column
from \code{colData} that includes the outcome variable.}

\item{valid.col}{\code{Character scalar}. Specifies the name of column
from \code{colData} that includes the information on which sample belongs to
validation set. The variable must be \code{logical}. (Default: \code{NULL})}

\item{...}{additional arguments passed to \code{IntegratedLearner()}
\itemize{
\item \code{na.rm}: \code{Logical scalar}. Should features with missing
values be dropped? (Default: \code{FALSE})
}}
}
\value{
A \code{SuperLearner} object containing the trained model fits.
}
\description{
This function offers a simple interface to the \code{IntegratedLearner()}
from a \code{\link{MultiAssayExperiment}} object. Instead of wrangling the
data, the function does it for the user and feeds it to
\code{IntegratedLearner()}.
}
