% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{NLIBD_MAE}
\alias{NLIBD_MAE}
\title{NLIBD binary-outcome MAE fixture}
\format{
A \code{MultiAssayExperiment} aligned to the PRISM MAE feature sets.
}
\usage{
data(NLIBD_MAE)
}
\description{
External validation cohort example data converted to
\code{MultiAssayExperiment} format for binary outcome modeling.
}
\keyword{datasets}
