% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_Utilities.R
\name{NNLS}
\alias{NNLS}
\title{NNLS function to optimize weights of several base learners}
\usage{
NNLS(x, y, wt)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{wt}{wt}
}
\value{
Solution of the quadratic programming problem
}
\description{
NNLS function to optimize weights of several base learners
}
\examples{
x <- cbind(c(0.1, 0.4, 0.6, 0.9), c(0.2, 0.5, 0.7, 0.8))
y <- c(0.15, 0.45, 0.65, 0.85)
fit <- NNLS(x = x, y = y, wt = rep(1, nrow(x)))
fit$solution
}
