% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_Utilities.R
\name{predict.SL.nnls.auc}
\alias{predict.SL.nnls.auc}
\title{Predict function for SL.nnls.auc}
\usage{
\method{predict}{SL.nnls.auc}(object, newdata, ...)
}
\arguments{
\item{object}{Fitted SL.nnls.auc model}

\item{newdata}{Validation layer-level predictions}

\item{...}{Additional arguments (unused)}
}
\value{
Prediction from the meta-learner
}
\description{
Predict function for SL.nnls.auc
}
