% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_Utilities.R
\name{predict.sl_nnls_auc}
\alias{predict.sl_nnls_auc}
\title{Predict function for sl_nnls_auc}
\usage{
\method{predict}{sl_nnls_auc}(object, newdata, ...)
}
\arguments{
\item{object}{Fitted sl_nnls_auc model}

\item{newdata}{Validation layer-level predictions}

\item{...}{Additional arguments (unused)}
}
\value{
Prediction from the meta-learner
}
\description{
Predict function for sl_nnls_auc
}
\examples{
set.seed(1)
X <- data.frame(m1 = runif(20), m2 = runif(20))
Y <- rnorm(20)
fit <- sl_nnls_auc(
  Y = Y, X = X, newX = X[1:4, ],
  family = stats::gaussian(), obsWeights = rep(1, nrow(X))
)
head(predict(fit$fit, newdata = X[1:4, ]))
}
