% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_Utilities.R
\name{auc.obj}
\alias{auc.obj}
\title{Title Meta level Objective function: NNLS for gaussian; Rank loss for binary observations}
\usage{
auc.obj(b, X, Y)
}
\arguments{
\item{b}{Weights vector}

\item{X}{Design matrix (data frame)}

\item{Y}{Outcome variable}
}
\value{
1 - AUC
}
\description{
Title Meta level Objective function: NNLS for gaussian; Rank loss for binary observations
}
