% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_Utilities.R
\name{NNLS}
\alias{NNLS}
\title{NNLS function to optimize weights of several base learners}
\usage{
NNLS(x, y, wt)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{wt}{wt}
}
\value{
Solution of the quadratic programming problem
}
\description{
NNLS function to optimize weights of several base learners
}
