% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLearner.R
\name{plot.learner}
\alias{plot.learner}
\title{Plot the summary curves produced by IntegratedLearner object}
\usage{
\method{plot}{learner}(fit, label_size = 8, label_x = 0.3, vjust = 0.1, rowwise_plot = TRUE)
}
\arguments{
\item{fit}{fitted "IntegratedLearner" object}

\item{label_size}{(optional) Numerical value indicating the label size. Default is 8.}

\item{label_x}{(optional) Single value or vector of x positions for plot labels, relative to each subplot. Defaults to 0.3 for all labels. (Each label is placed all the way to the left of each plot.)}

\item{vjust}{Adjusts the vertical position of each label. More positive values move the label further down on the plot canvas. Can be a single value (applied to all labels) or a vector of values (one for each label). Default is 0.1.}

\item{rowwise_plot}{If both train and test data is available, should the train and test plots be rowwise_plot. Default is TRUE. If FALSE, plots are aligned column-wise.}
}
\value{
ggplot2 object
}
\description{
Plots the R^2/AUC curves for the training (and test, if provided) set produced by IntegratedLearner object
}
