% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_Utilities.R
\name{predict.SL.BART}
\alias{predict.SL.BART}
\title{Predict function for SL.BART}
\usage{
\method{predict}{SL.BART}(object, newdata, family, X = NULL, Y = NULL, ...)
}
\arguments{
\item{object}{object}

\item{newdata}{newdata}
}
\value{
Prediction from the SL.BART
}
\description{
Predict function for SL.BART
}
