% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_Utilities.R
\name{SL.horseshoe}
\alias{SL.horseshoe}
\title{Horseshoe regression}
\usage{
SL.horseshoe(
  Y,
  X,
  newX,
  family,
  prior = "horseshoe",
  N = 20000L,
  burnin = 1000L,
  thinning = 1L,
  ...
)
}
\arguments{
\item{Y}{Outcome variable}

\item{X}{Covariate data frame}

\item{newX}{Dataframe to predict the outcome}

\item{family}{"gaussian" for regression, "binomial" for binary
classification. Untested options: "poisson" for for integer or count data}

\item{prior}{prior for regression coefficients to use. "Horseshoe" by default. 
Untested options: ridge regression (prior="rr" or prior="ridge"), 
lasso regression (prior="lasso") and horseshoe+ regression (prior="hs+" 
or prior="horseshoe+")}

\item{N}{Number of posterior samples to generate.}

\item{burnin}{Number of burn-in samples.}

\item{thinning}{Desired level of thinning.}

\item{...}{other parameters passed to bayesreg function}
}
\value{
SL object
}
\description{
Horseshoe regression
}
