% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_Utilities.R
\name{SL.nnls.auc}
\alias{SL.nnls.auc}
\title{Combined SuperLearner function for both NNLS/AUC maximization}
\usage{
SL.nnls.auc(Y, X, newX, family, obsWeights, bounds = c(0, Inf), ...)
}
\arguments{
\item{Y}{Outcome matrix from metalearner}

\item{X}{Layer-level predictions used to train the metalearner}

\item{newX}{Layer-level predictions for validation data}

\item{family}{Family object}

\item{obsWeights}{Observation weights}

\item{bounds}{Lower/upper bounds for weights (binomial case)}

\item{...}{Additional arguments passed through}
}
\value{
Estimated meta-learner coefficients and predictions
}
\description{
Combined SuperLearner function for both NNLS/AUC maximization
}
