% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_Utilities.R
\name{SL.nnls.auc}
\alias{SL.nnls.auc}
\title{Combined SuperLearner function for both NNLS/AUC maximization}
\usage{
SL.nnls.auc(Y, X, newX, family, obsWeights, bounds = c(0, Inf), ...)
}
\arguments{
\item{Y}{Y}

\item{X}{X}
}
\value{
Estimated meta-learner coefficients and predictions
}
\description{
Combined SuperLearner function for both NNLS/AUC maximization
}
