% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLearner.R
\name{update.learner}
\alias{update.learner}
\title{Update IntegratedLearner fit object based on layers available in the test set}
\usage{
update.learner(
  fit,
  feature_table_valid,
  sample_metadata_valid = NULL,
  feature_metadata_valid,
  seed = 1234,
  verbose = FALSE
)
}
\arguments{
\item{fit}{fitted "IntegratedLearner" object}

\item{feature_table_valid}{Feature table from validation set. It should be a data frame with features in rows and samples in columns. Feature names should be a subset of training data feature names.}

\item{sample_metadata_valid}{OPTIONAL (can provide feature_table_valid and not this):  Sample-specific metadata table from independent validation set. If provided, it must have the exact same structure as sample_metadata. Default is NULL.}

\item{feature_metadata_valid}{Matrix containing feature names and their corresponding layers. Must be subset of feature_metadata provided in IntegratedLearner object.}

\item{seed}{Seed for reproducibility. Default is 1234.}

\item{verbose}{Should a summary of fits/ results be printed. Default is FALSE}
}
\value{
SL object
}
\description{
Allow update of IntegratedLearner if only a subset of omics layers are available in test set. If all layers and features match, it calls predict.learner()
}
